"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2 Assignments/Module 6/m6/ADS-2Week-1/Solution.java
@@ -24,7 +24,7 @@ class PageRank {
 		System.out.println(v + "v'svalue");
 		for(Integer eachadj : pggraph.adj(v)) {
 			System.out.println(prval[eachadj]/(pggraph.outdegree(eachadj)) + "everytime" + eachadj);
-			testprval = testprval + ((prval[eachadj])/(pggraph.outdegree(eachadj)));
+			testprval = testprval + ((prval[eachadj])/(double)(pggraph.outdegree(eachadj)));
 		}
 		System.out.println(testprval);
 		prval[v] = testprval;
